Try to attain fixed size of kind icon container on all platforms
The goal is to have a 25x25 pixel size (with 100% UI scaling) for the
kind icon container on all platforms, so that the icon background
texture fits into the kind container. This requires small adjustments to
the font size and/or margin of the kind letter. The following fonts are
assumed for the proportional "system" font:
  - Windows: Segoe UI
  - macOS: San Francisco
  - Linux: Ubuntu