Improve hovered tab rendering with "angled" tab style
Prevents tab overlaps to be visible for the half-transparent hovered tabs.
There still seems to be a bug with the "right_of_hovered" attribute
(ST issue 6238) sometimes. And "!right_of_selected" should actually
prevent this bug to be visible, but it doesn't seem to have any effect
here. I might look into that later.