Fix syntax specific settings opening the wrong settings file
The 'Preferences: Settings - Syntax Specific' command from the command
palette or the menu uses the syntax filename to open the corresponding
syntax specific settings file. The syntax aliases which are needed for
the special file icons did have an additional label in the filename,
so using the command would open a settings file with the according
filename, instead of the regular settings file for the original syntax.
This commit adjust the filenames of the syntax aliases in accordance to
their derived syntax names. Similarly, the label suffix in the display
name is also removed.